feat(RNNModel): renamed 'RNNModel' to 'BlockRNNModel' and 'TrueRNNModel' to 'RNNModel'
