+++ darts/models/rnn_model.py
@@ -5,26 +5,29 @@ Recurrent Neural Networks
 
 import torch.nn as nn
 import torch
+from torch.utils.data import DataLoader
 from numpy.random import RandomState
-from typing import List, Optional, Union
+from joblib import Parallel, delayed
+from typing import Sequence, Optional, Union
+from ..timeseries import TimeSeries
 
 from ..logging import raise_if_not, get_logger
+from .torch_forecasting_model import TorchForecastingModel, TimeSeriesTorchDataset
 from ..utils.torch import random_method
-from .torch_forecasting_model import TorchForecastingModel
+from ..utils.data.timeseries_dataset import TimeSeriesInferenceDataset
+from ..utils.data import ShiftedDataset
+from ..utils import _build_tqdm_iterator
 
 logger = get_logger(__name__)
 
 
 # TODO add batch norm
-class _RNNModule(nn.Module):
+class _TrueRNNModule(nn.Module):
     def __init__(self,
                  name: str,
                  input_size: int,
                  hidden_dim: int,
-                 num_layers: int,
-                 output_chunk_length: int = 1,
                  target_size: int = 1,
-                 num_layers_out_fc: Optional[List] = None,
                  dropout: float = 0.):
 
         """ PyTorch module implementing a RNN to be used in `RNNModel`.
@@ -42,10 +45,6 @@ class _RNNModule(nn.Module):
             The dimensionality of the input time series.
         hidden_dim
             The number of features in the hidden state `h` of the RNN module.
-        num_layers
-            The number of recurrent layers.
-        output_chunk_length
-            The number of steps to predict in the future.
         target_size
             The dimensionality of the output time series.
         num_layers_out_fc
@@ -56,65 +55,55 @@ class _RNNModule(nn.Module):
 
         Inputs
         ------
-        x of shape `(batch_size, input_chunk_length, input_size)`
-            Tensor containing the features of the input sequence.
+        x of shape `(batch_size, input_length, input_size)`
+            Tensor containing the features of the input sequence. The `input_length` is not fixed.
 
         Outputs
         -------
-        y of shape `(batch_size, out_len, output_size)`
-            Tensor containing the (point) prediction at the last time step of the sequence.
+        y of shape `(batch_size, output_length, output_size)`
+            The `output_length` is equal to 1 at prediction time, but equal to the `input_length` during training.
         """
 
-        super(_RNNModule, self).__init__()
+        super(_TrueRNNModule, self).__init__()
 
         # Defining parameters
-        self.hidden_dim = hidden_dim
-        self.n_layers = num_layers
         self.target_size = target_size
-        num_layers_out_fc = [] if num_layers_out_fc is None else num_layers_out_fc
-        self.out_len = output_chunk_length
         self.name = name
 
         # Defining the RNN module
-        self.rnn = getattr(nn, name)(input_size, hidden_dim, num_layers, batch_first=True, dropout=dropout)
+        self.rnn = getattr(nn, name)(input_size, hidden_dim, 1, batch_first=True, dropout=dropout)
 
-        # The RNN module is followed by a fully connected layer, which maps the last hidden layer
-        # to the output of desired length
-        last = hidden_dim
-        feats = []
-        for feature in num_layers_out_fc + [output_chunk_length * target_size]:
-            feats.append(nn.Linear(last, feature))
-            last = feature
-        self.fc = nn.Sequential(*feats)
+        # The RNN module needs a linear layer V that transforms hidden states into outputs, individually
+        self.V = nn.Linear(hidden_dim, target_size)
 
-    def forward(self, x):
-        # data is of size (batch_size, input_chunk_length, input_size)
+    def forward(self, x, h=None):
+        # data is of size (batch_size, input_length, input_size)
         batch_size = x.size(0)
 
-        out, hidden = self.rnn(x)
+        # out is of size (batch_size, input_length, hidden_dim)
+        out, last_hidden_state = self.rnn(x) if h is None else self.rnn(x, h)
+        # TODO: confirm shape of out
 
-        """ Here, we apply the FC network only on the last output point (at the last time step)
+        """ Here, we apply the V matrix to every hidden state to produce the outputs
         """
-        if self.name == "LSTM":
-            hidden = hidden[0]
-        predictions = hidden[-1, :, :]
-        predictions = self.fc(predictions)
-        predictions = predictions.view(batch_size, self.out_len, self.target_size)
+        # TODO: for one layer out should be equal to hidden, make sure that's true
+        predictions = self.V(out)  # TODO: make sure this matrix multiplication is correct
 
-        # predictions is of size (batch_size, output_chunk_length, 1)
-        return predictions
+        # predictions is of size (batch_size, input_length, target_size)
+        predictions = predictions.view(batch_size, -1, self.target_size)
+
+        # returns outputs for all inputs, only the last one is needed for prediction time
+        return predictions, last_hidden_state
 
 
 class RNNModel(TorchForecastingModel):
     @random_method
     def __init__(self,
-                 input_chunk_length: int,
-                 output_chunk_length: int,
                  model: Union[str, nn.Module] = 'RNN',
-                 hidden_size: int = 25,
-                 n_rnn_layers: int = 1,
-                 hidden_fc_sizes: Optional[List] = None,
+                 input_chunk_length: int = 12,
+                 hidden_dim: int = 25,
                  dropout: float = 0.,
+                 training_length: int = 24,
                  random_state: Optional[Union[int, RandomState]] = None,
                  **kwargs):
 
@@ -135,24 +124,20 @@ class RNNModel(TorchForecastingModel):
             or a PyTorch module with the same specifications as
             `darts.models.rnn_model.RNNModule`.
         input_chunk_length
-            The number of time steps that will be fed to the internal forecasting module
-        output_chunk_length
-            Number of time steps to be output by the internal forecasting module.
-        hidden_size
+            Number of past time steps that are fed to the forecasting module at prediction time.
+        hidden_dim
             Size for feature maps for each hidden RNN layer (:math:`h_n`).
-        n_rnn_layers
-            Number of layers in the RNN module.
-        hidden_fc_sizes
-            Sizes of hidden layers connecting the last hidden layer of the RNN module to the output, if any.
         dropout
             Fraction of neurons afected by Dropout.
+        training_length
+            The length of 1 training sample time series.
         random_state
             Control the randomness of the weights initialization. Check this
             `link <https://scikit-learn.org/stable/glossary.html#term-random-state>`_ for more details.
         """
 
         kwargs['input_chunk_length'] = input_chunk_length
-        kwargs['output_chunk_length'] = output_chunk_length
+        kwargs['output_chunk_length'] = 1
         super().__init__(**kwargs)
 
         # check we got right model type specified:
@@ -161,25 +146,31 @@ class RNNModel(TorchForecastingModel):
                                                        '"GRU", or give your own PyTorch nn.Module'.format(
                                                         model.__class__.__name__), logger)
 
-        self.input_chunk_length = input_chunk_length
-        self.output_chunk_length = output_chunk_length
         self.rnn_type_or_module = model
-        self.hidden_fc_sizes = hidden_fc_sizes
-        self.hidden_size = hidden_size
-        self.n_rnn_layers = n_rnn_layers
         self.dropout = dropout
+        self.hidden_dim = hidden_dim
+        self.training_length = training_length
+        self.is_recurrent = True
 
     def _create_model(self, input_dim: int, output_dim: int) -> torch.nn.Module:
         if self.rnn_type_or_module in ['RNN', 'LSTM', 'GRU']:
-            hidden_fc_sizes = [] if self.hidden_fc_sizes is None else self.hidden_fc_sizes
-            model = _RNNModule(name=self.rnn_type_or_module,
-                               input_size=input_dim,
-                               target_size=output_dim,
-                               hidden_dim=self.hidden_size,
-                               num_layers=self.n_rnn_layers,
-                               output_chunk_length=self.output_chunk_length,
-                               num_layers_out_fc=hidden_fc_sizes,
-                               dropout=self.dropout)
+            model = _TrueRNNModule(name=self.rnn_type_or_module,
+                                   input_size=input_dim,
+                                   target_size=output_dim,
+                                   hidden_dim=self.hidden_dim,
+                                   dropout=self.dropout)
         else:
             model = self.rnn_type_or_module
         return model
+
+    def _build_train_dataset(self,
+                             target: Sequence[TimeSeries],
+                             covariates: Optional[Sequence[TimeSeries]]) -> ShiftedDataset:
+        return ShiftedDataset(target_series=target,
+                              covariates=covariates,
+                              length=self.training_length,
+                              shift_covariates=True,
+                              shift=1)
+
+    def _produce_train_output(self, data):
+        return self.model(data)[0]
