+++ darts/tests/test_regression_ensemble_model.py
@@ -12,7 +12,7 @@ from ..logging import get_logger
 logger = get_logger(__name__)
 
 try:
-    from ..models import RNNModel
+    from ..models import BlockRNNModel
     TORCH_AVAILABLE = True
 except ImportError:
     logger.warning('Torch not available. Some tests will be skipped.')
@@ -78,8 +78,8 @@ class RegressionEnsembleModelsTestCase(DartsBaseTestClass):
 
     if TORCH_AVAILABLE:
         def test_torch_models_retrain(self):
-            model1 = RNNModel(input_chunk_length=12, output_chunk_length=1, random_state=0, n_epochs=2)
-            model2 = RNNModel(input_chunk_length=12, output_chunk_length=1, random_state=0, n_epochs=2)
+            model1 = BlockRNNModel(input_chunk_length=12, output_chunk_length=1, random_state=0, n_epochs=2)
+            model2 = BlockRNNModel(input_chunk_length=12, output_chunk_length=1, random_state=0, n_epochs=2)
 
             ensemble = RegressionEnsembleModel([model1], 5)
             ensemble.fit(self.combined)
