+++ darts/utils/data/simple_inference_dataset.py
@@ -35,9 +35,9 @@ class SimpleInferenceDataset(TimeSeriesInferenceDataset):
         Recurrent models:
         If covariates were used to train the model, `n` covariates have to be available into the future.
         Therefore, `cov_future` will be equal to the future covariates up to `n` time steps into the future.
+
         The parameter `input_chunk_length` is necessary to determine the minimum length for all `tgt_past`
         and `cov_past` time series.
-
         Parameters `n`, `output_chunk_length` and `model_is_recurrent` are necessary to determine whether
         `cov_future` is necessary (or can be set to `None`) and to determine the required length for `cov_future`.
         It is important for the 3 emitted time series to have the same length across data points because
